поправлены замечания. НЕ ВСЕ
