+++ spring-6/src/main/java/ru/otus/hmwrk/dao/AuthorsRepository.java
@@ -2,16 +2,13 @@ package ru.otus.hmwrk.dao;
 
 import jakarta.persistence.EntityGraph;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.otus.hmwrk.entity.Author;
-import ru.otus.hmwrk.exceptions.AuthorNameNotValidException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -41,20 +38,6 @@ public class AuthorsRepository implements Repository<Author, Long> {
                 Author.class);
         query.setParameter("firstName", firstName);
         return query.getResultList();
-    }
-
-    @Override
-    public void updateNameById(Long id, String name) {
-        var names = name.split(" ");
-        if (names.length == 2) {
-            var author = Optional.ofNullable(findById(id))
-                    .orElseThrow(() -> new EntityNotFoundException("Entity with id: " + id + "not found"));
-            author.setFirstName(names[0]);
-            author.setLastName(names[1]);
-            em.merge(author);
-        } else {
-            throw new AuthorNameNotValidException(name);
-        }
     }
 
     @Override
