+++ spring-6/src/main/java/ru/otus/hmwrk/dao/BooksRepository.java
@@ -1,19 +1,14 @@
 package ru.otus.hmwrk.dao;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Component;
 import ru.otus.hmwrk.entity.Book;
-import ru.otus.hmwrk.exceptions.NotValidIdentifierException;
 
 import java.util.List;
-import java.util.Optional;
-
-import static java.util.Objects.nonNull;
 
 
 /**
@@ -46,19 +41,6 @@ public class BooksRepository implements Repository<Book, Long> {
                 Book.class);
         query.setParameter("name", name);
         return query.getResultList();
-    }
-
-    @Override
-    public void updateNameById(Long id, String name) {
-        if (nonNull(id)){
-            var book = Optional.ofNullable(findById(id))
-                    .orElseThrow(() -> new EntityNotFoundException("Entity with id: " + id + "not found"));
-
-            book.setName(name);
-            em.merge(book);
-        } else {
-            throw new NotValidIdentifierException(id);
-        }
     }
 
     @Override
