+++ spring-6/src/main/java/ru/otus/hmwrk/dao/Repository.java
@@ -4,11 +4,10 @@ import java.util.List;
 
 public interface Repository<T, ID> {
     List<T> findAll();
+
     T findById(ID id);
 
     List<T> findByName(String name);
-
-    void updateNameById(ID id, String name);
 
     T save(T entity);
 }
