+++ spring-6/src/main/java/ru/otus/hmwrk/service/BookService.java
@@ -1,13 +1,17 @@
 package ru.otus.hmwrk.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.hmwrk.dao.BooksRepository;
 import ru.otus.hmwrk.entity.Book;
+import ru.otus.hmwrk.exceptions.NotValidIdentifierException;
 
 import java.util.List;
 import java.util.Optional;
+
+import static java.util.Objects.nonNull;
 
 @Service
 @RequiredArgsConstructor
@@ -17,18 +21,28 @@ public class BookService implements CommonService<Book, Long> {
 
     @Override
     public List<Book> findAll() {
-       return booksRepository.findAll();
+        return booksRepository.findAll();
     }
 
     @Override
-    @Transactional(readOnly = true)
     public Optional<Book> findById(Long id) {
         return Optional.ofNullable(booksRepository.findById(id));
     }
 
     @Override
-    @Transactional
     public Optional<Book> save(Book entity) {
         return Optional.ofNullable(booksRepository.save(entity));
+    }
+
+    @Override
+    public void updateNameById(Long id, String name) {
+        if (nonNull(id)) {
+            var book = findById(id)
+                    .orElseThrow(() -> new EntityNotFoundException("Entity with id: " + id + "not found"));
+            book.setName(name);
+            save(book);
+        } else {
+            throw new NotValidIdentifierException(id);
+        }
     }
 }
